refactor(elements): define base-element
